A few fixes to atsmake
In general, the construction:

ifeq ("$(VAR)","VALUE")
...

doesn't work as expected; expansion of variables with whitespace can
cause problems. The solution is to do:

ifeq ($(strip $(VAR),VALUE))
...

which strips off any extra whitespace.

In the case where VALUE is the empty string, it makes more sense to
write:

ifndef VAR
...

Which results in many instances of the construction:

ifndef ...
else
...
endif

I changed this to:

ifdef
...
endif

I also removed the initial empty declarations of SOURCES_?ATS in
atsmake-pre.mk; This allows for the use of ifdef.

for more info, see:

https://gnu.org/software/make/manual/make.html#Conditional-Syntax